requirements.txt
Flask>=2.3.0,<3.0.0
pywebview>=4.4
